moved masonry options directly into portfolio container file. changed filterbuttons to be in filterbuttons folder
