--- src/components/Portfolio/Exports.jsx
@@ -1,5 +1,0 @@
-export const masonryOptions = {
-  fitWidth: true,
-  columnWidth: 300,
-  gutter: 5
-};
